+++ src/api/objects/toot.ts
@@ -427,7 +427,7 @@ export default class Toot implements TootObj {
 
     /**
      * Return all but the last paragraph if that last paragraph is just hashtag links.
-     * @param {number} [fontSize=DEFAULT_FONT_SIZE] - Size in pixels to render emojis as <img> tags. Should match surrounding txt..
+     * @param {number} [fontSize=DEFAULT_FONT_SIZE] - Size in pixels of any emoji <img> tags. Should match surrounding txt.
      * @returns {string}
      */
     contentNonTagsParagraphs(fontSize: number = DEFAULT_FONT_SIZE): string {
@@ -438,7 +438,7 @@ export default class Toot implements TootObj {
 
     /**
      * Break up the content into paragraphs and add <img> tags for custom emojis.
-     * @param {number} [fontSize=DEFAULT_FONT_SIZE] - Size in pixels to render emojis as <img> tags. Should match surrounding txt..
+     * @param {number} [fontSize=DEFAULT_FONT_SIZE] - Size in pixels of any emoji <img> tags. Should match surrounding txt.
      * @returns {string[]}
      */
     contentParagraphs(fontSize: number = DEFAULT_FONT_SIZE): string[] {
@@ -466,7 +466,7 @@ export default class Toot implements TootObj {
 
     /**
      * Replace custom emoji shortcodes (e.g. ":myemoji:") with image tags.
-     * @param {number} [fontSize=DEFAULT_FONT_SIZE] - Size in pixels to render emoji imgs. Should match surrounding text.
+     * @param {number} [fontSize=DEFAULT_FONT_SIZE] - Size in pixels of any emoji imgs. Should match surrounding text.
      * @returns {string}
      */
     contentWithEmojis(fontSize: number = DEFAULT_FONT_SIZE): string {