+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # NEXT RELEASE
+* **(BREAKING CHANGE)** Add `fontSize` argument to `Account.noteWithAccountInfo()`, which is now a fxn not a getter
+* (Demo App) Add `config.theme.accountNoteFontSize` property
 
 ### v1.1.3
 * Export `DEFAULT_FONT_SIZE`